# Conflicts:
#	sec/webapp/controllers/hypothesis.go

+++ sec/library/models/powerplantcoordinates.go
@@ -19,14 +19,14 @@ type PowerPlantCoordinates struct {
 	FuelTypes_Heavy       bool    `bson:"FuelTypes_Heavy" json:"FuelTypes_Heavy"`
 	FuelTypes_Diesel      bool    `bson:"FuelTypes_Diesel" json:"FuelTypes_Diesel"`
 	FuelTypes_Gas         bool    `bson:"FuelTypes_Gas" json:"FuelTypes_Gas"`
-	GasTurbineUnit        float64 `bson:"GasTurbineUnit" json:"GasTurbineUnit"`
-	GasTurbineCapacity    float64 `bson:"GasTurbineCapacity" json:"GasTurbineCapacity"`
+	GasTurbineUnit        int     `bson:"GasTurbineUnit" json:"GasTurbineUnit"`
+	GasTurbineCapacity    int     `bson:"GasTurbineCapacity" json:"GasTurbineCapacity"`
 	SteamUnit             int     `bson:"SteamUnit" json:"SteamUnit"`
-	SteamCapacity         float64 `bson:"SteamCapacity" json:"SteamCapacity"`
+	SteamCapacity         int     `bson:"SteamCapacity" json:"SteamCapacity"`
 	DieselUnit            int     `bson:"DieselUnit" json:"DieselUnit"`
-	DieselCapacity        float64 `bson:"DieselCapacity" json:"DieselCapacity"`
+	DieselCapacity        int     `bson:"DieselCapacity" json:"DieselCapacity"`
 	CombinedCycleUnit     int     `bson:"CombinedCycleUnit" json:"CombinedCycleUnit"`
-	CombinedCycleCapacity float64 `bson:"CombinedCycleCapacity" json:"CombinedCycleCapacity"`
+	CombinedCycleCapacity int     `bson:"CombinedCycleCapacity" json:"CombinedCycleCapacity"`
 	Longitude             float64 `bson:"Longitude" json:"Longitude"`
 	Latitude              float64 `bson:"Latitude" json:"Latitude"`
 }
